Bag in album fixed, any fixes

+++ src/services/products.ts
@@ -1,4 +1,4 @@
-import { Product } from "../types/products";
+import { GetProductsParams, Product } from "../types/products";
 import {
   createProduct as dbCreateProduct,
   getProducts as dbGetProducts,
@@ -6,22 +6,22 @@ import {
   updateProduct as dbUpdateProduct,
 } from "../database/repositories/products";
 import fs from "fs/promises";
-
 import { saveImage, saveAlbum, getImgPath } from "../services/uploadService";
 import ShortUniqueId from "short-unique-id";
+import { RepositoryResponse } from "../types/repositoryResponse";
 
 const uid = new ShortUniqueId({ length: 10 });
 
-interface GetProductsParams {
-  title?: string;
-  sortByPrice?: "asc" | "desc";
-  page?: number;
-  limit?: number | "*";
-}
-
 const productService = {
-  getProducts: async (params: GetProductsParams) => {
-    return dbGetProducts(params);
+  getProducts: async (
+    params: GetProductsParams
+  ): Promise<RepositoryResponse<Product[]>> => {
+    const response = await dbGetProducts(params);
+    return {
+      data: response.data,
+      errorMessage: response.errorMessage,
+      errorRaw: response.errorRaw,
+    };
   },
 
   createProduct: async (
@@ -29,10 +29,10 @@ const productService = {
     amount: number,
     price: number,
     favorite: boolean,
-    image: any,
-    albumPhotos: any[]
-  ) => {
-    const newProduct: any = {
+    image: { name: string; data: Buffer } | null,
+    albumPhotos: { name: string; data: Buffer }[]
+  ): Promise<RepositoryResponse<Product>> => {
+    const newProduct: Product = {
       id: uid.rnd(),
       title,
       amount,
@@ -45,42 +45,61 @@ const productService = {
           : [],
     };
 
-    return dbCreateProduct(newProduct);
+    const response = await dbCreateProduct(newProduct);
+    return {
+      data: response.data,
+      errorMessage: response.errorMessage,
+      errorRaw: response.errorRaw,
+    };
   },
 
-  editTitle: async (productId: string, updatedData: Partial<Product>) => {
-    return dbUpdateProduct(productId, updatedData);
+  editTitle: async (
+    productId: string,
+    updatedData: Partial<Product>
+  ): Promise<RepositoryResponse<Product>> => {
+    const response = await dbUpdateProduct(productId, updatedData);
+    return {
+      data: response.data,
+      errorMessage: response.errorMessage,
+      errorRaw: response.errorRaw,
+    };
   },
 
-  deleteProduct: async (productId: string) => {
+  deleteProduct: async (
+    productId: string
+  ): Promise<RepositoryResponse<Product>> => {
     const deletedProduct = await dbDeleteProduct(productId);
 
     if (deletedProduct.data) {
-      try {
-        if (deletedProduct.data.image) {
-          const imagePath = getImgPath(deletedProduct.data.image);
+      if (deletedProduct.data.image) {
+        const imagePath = getImgPath(deletedProduct.data.image);
+        if (imagePath) {
           await fs.unlink(imagePath);
         }
+      }
 
-        if (
-          deletedProduct.data.albumPhotos &&
-          deletedProduct.data.albumPhotos.length > 0
-        ) {
-          const deletePromises = deletedProduct.data.albumPhotos.map(
-            async (photo: string) => {
-              const photoPath = getImgPath(photo);
+      if (
+        deletedProduct.data.albumPhotos &&
+        deletedProduct.data.albumPhotos.length > 0
+      ) {
+        const deletePromises = deletedProduct.data.albumPhotos.map(
+          async (photo) => {
+            const photoPath = getImgPath(photo);
+            if (photoPath) {
               return fs.unlink(photoPath);
             }
-          );
+          }
+        );
 
-          await Promise.all(deletePromises);
-        }
-      } catch (error) {
-        console.error("Failed to delete product images:", error);
+        await Promise.all(deletePromises);
       }
     }
 
-    return deletedProduct;
+    return {
+      data: deletedProduct.data,
+      errorMessage: deletedProduct.errorMessage,
+      errorRaw: deletedProduct.errorRaw,
+    };
   },
 };
